requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.14.0
Pillow>=9.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
opencv-python>=4.8.0
pycocotools>=2.0.6
tqdm>=4.65.0
wandb>=0.15.0
timm>=0.9.0
peft>=0.7.0
streamlit>=1.28.0
